fix double-free
when no scaling was required, the image would be freed twice. Fix it
by always allocating a new image, even when it would not otherwise be
required. The alternative would be to require that the client always
tests that a new image was allocated, but that is a really terrible
interface.